Move Payment to the daemon package
