+++ .gitattributes
@@ -1,0 +1,1 @@
+*.svg binary
