+++ src/main/java/io/singularitynet/sdk/daemon/PaymentSerializer.java
@@ -1,10 +1,19 @@
-package io.singularitynet.sdk.mpe;
+package io.singularitynet.sdk.daemon;
 
 import io.grpc.Metadata;
 import java.util.Optional;
+import java.util.Map;
+import java.util.HashMap;
 import java.math.BigInteger;
+import java.util.function.Function;
 
 public class PaymentSerializer {
+
+    private static final Map<String, Function<Metadata, Payment>> readerByType = new HashMap<>();
+
+    public static void register(String type, Function<Metadata, Payment> reader) {
+        readerByType.put(type, reader);
+    }
 
     public static Optional<Payment> fromMetadata(Metadata headers) {
         if (!headers.containsKey(Payment.SNET_PAYMENT_TYPE)) {
@@ -12,18 +21,20 @@ public class PaymentSerializer {
         }
 
         String paymentType = headers.get(Payment.SNET_PAYMENT_TYPE);
-        if (paymentType.equals(EscrowPayment.PAYMENT_TYPE_ESCROW)) {
-            return Optional.of(EscrowPayment.fromMetadata(headers)); 
-        } else {
+        Function<Metadata, Payment> reader = readerByType.get(paymentType);
+
+        if (reader == null) {
             throw new IllegalArgumentException("Unexpected payment type: " + paymentType);
         }
+
+        return Optional.of(reader.apply(headers)); 
     }
 
     public static void toMetadata(Payment payment, Metadata headers) {
         payment.toMetadata(headers);
     }
 
-    static final Metadata.AsciiMarshaller<BigInteger> ASCII_BIGINTEGER_MARSHALLER =
+    public static final Metadata.AsciiMarshaller<BigInteger> ASCII_BIGINTEGER_MARSHALLER =
         new Metadata.AsciiMarshaller<BigInteger>() {
 
             @Override
